refactor meta functions
